+++ backend/analysis/annotations/safreader_test.py
@@ -1,9 +1,10 @@
-from operator import itemgetter
-import pytest
 import os.path as op
+from operator import itemgetter
 
+import pytest
+from analysis.annotations.safreader import SAFReader, get_word_levels
 from analysis.query.run import query_transcript
-from analysis.annotations.safreader import SAFReader
+from pandas import DataFrame
 from pytest_lazyfixture import lazy_fixture
 
 
@@ -29,10 +30,15 @@ def test_read_saf(method, transcript, filedir, samplenum):
     assert true_results.annotations.keys() == reader.document.reformatted_annotations.keys()
     for q, annos in true_results.annotations.items():
         true_annos = reader.document.reformatted_annotations[q]
-        for word, true_word in zip(annos, true_annos):
+        for word in annos:
+            true_word = next((w for w in true_annos if w.index == word.index), None)
             hits = sorted(word.hits, key=itemgetter('level', 'item'))
-            true_hits = sorted(true_word.hits, key=itemgetter('level', 'item'))
-            assert hits == true_hits
+            if true_word:
+                true_hits = sorted(true_word.hits, key=itemgetter('level', 'item'))
+                assert hits == true_hits
+            else:
+                # if the true_word is not found (unaligned empty), make sure it didnt miss anything
+                assert hits == []
 
     # are the exactresults the same?
     true_exact = {k: sorted(v) for (k, v) in true_results.exactresults.items() if v != []}
@@ -50,3 +56,20 @@ def test_astalex(asta_method, asta_transcript, asta_transcript_corrections, cha_
     assert true_results.annotations.get(3)[6].hits == [{'level': 'Taalmaat', 'item': 'N', 'fase': None}]
 
     assert True
+
+
+def test_wordlevels():
+    data = {'level': map(str.lower, ['Utt', 'QA', 'SZ', 'Grammaticale Fout', 'Commentaar']),
+            'word1': [1, None, 'X', 'V, BvBB', 'Hier staat wat commentaar']}
+    df_in = DataFrame.from_dict(data)
+
+    word_levels = get_word_levels(df_in)
+    assert word_levels == ['qa', 'sz', 'grammaticale fout']
+
+
+def test_read_saf_comments(tarsp_method, tarsp_transcript, cha_testfiles_dir):
+    reader = SAFReader(op.join(cha_testfiles_dir, 'sample_5_SAF_with_comments.xlsx'), tarsp_method, tarsp_transcript)
+    sent = reader.document.utterances[3]
+    assert sent.words[0].comment == 'Ik vind hier iets van.'
+    assert sent.words[1].comment == '1'
+    assert sent.words[2].comment == 'En hier misschien ook wel iets van'