+++ backend/analysis/views.py
@@ -8,10 +8,9 @@ from io import BytesIO, StringIO
 from analysis.annotations.enrich_chat import enrich_chat
 from analysis.annotations.safreader import SAFReader
 from analysis.query.run import query_transcript
-from analysis.query.xlsx_output import v1_to_xlsx, v2_to_xlsx
+from analysis.query.xlsx_output import annotations_to_xlsx, querycounts_to_xlsx
 from celery import group
 from convert.chat_writer import ChatWriter
-from django.conf import settings
 from django.db.models import Q
 from django.http import HttpResponse
 from parse.parse_utils import parse_and_create
@@ -21,8 +20,6 @@ from rest_framework.decorators import action
 from rest_framework.exceptions import ParseError
 from rest_framework.response import Response
 
-from sasta.celery import get_celery_worker_status
-
 from .convert.convert import convert
 from .models import (AnalysisRun, AssessmentMethod, Corpus, MethodCategory,
                      Transcript, UploadFile)
@@ -31,9 +28,6 @@ from .serializers import (AssessmentMethodSerializer, CorpusSerializer,
                           MethodCategorySerializer, TranscriptSerializer,
                           UploadFileSerializer)
 from .utils import StreamFile
-
-logger = logging.getLogger('sasta')
-
 
 # flake8: noqa: E501
 SPREADSHEET_MIMETYPE = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
@@ -86,7 +80,7 @@ class TranscriptViewSet(viewsets.ModelViewSet):
 
         allresults, queries_with_funcs = query_transcript(transcript, method)
 
-        spreadsheet = v1_to_xlsx(allresults, queries_with_funcs)
+        spreadsheet = querycounts_to_xlsx(allresults, queries_with_funcs)
         spreadsheet.save(response)
 
         return response
@@ -103,17 +97,14 @@ class TranscriptViewSet(viewsets.ModelViewSet):
             transcript, method, True, zc_embed
         )
 
-        spreadsheet = v2_to_xlsx(allresults, method, zc_embeddings=zc_embed)
+        spreadsheet = annotations_to_xlsx(allresults, method)
         self.create_analysis_run(transcript, method, spreadsheet)
 
         format = request.data.get('format', 'xlsx')
 
         if format == 'xlsx':
-            spreadsheet = v2_to_xlsx(
-                allresults, method, zc_embeddings=zc_embed)
-
             response = HttpResponse(
-                content_type='application/vnd.openxmlformats-officedocument.spreadsheetml.sheet')
+                content_type=SPREADSHEET_MIMETYPE)
             response['Content-Disposition'] = "attachment; filename=saf_output.xlsx"
             spreadsheet.save(response)